Restructure generation of settings data feed items to be usable from custom generators and implement subgroups
