+++ MonkeyLoader.Resonite.Integration/Locale/FallbackLocaleGenerationEvent.cs
@@ -1,12 +1,7 @@
 ﻿using Elements.Assets;
 using EnumerableToolkit;
 using MonkeyLoader.Events;
-using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MonkeyLoader.Resonite.Locale
 {
